commands/status: Set success_message / failure_message to fallback field
Fixes #47

In the current implementation, using `failure_message` in status doesn't
affect the `fallback` field.
It causes a problem when sending a message to bots such as `hubot-slack`.
The `hubot-slack` parses the `fallback` field if a message has an attachment.

We should set `text` and `fallback` field to the same value.
The `success_message` also have the same issue. So I fixed both.